feat: migrate workspace access control to separate crate
