+++ Cargo.toml
@@ -89,6 +89,7 @@ gotrue = { path = "libs/gotrue" }
 gotrue-entity = { path = "libs/gotrue-entity" }
 infra = { path = "libs/infra" }
 access-control.workspace = true
+workspace-access.workspace = true
 app-error = { workspace = true, features = ["sqlx_error", "actix_web_error", "tokio_error"] }
 shared-entity = { path = "libs/shared-entity", features = ["cloud"] }
 workspace-template = { workspace = true }
@@ -137,6 +138,7 @@ members = [
     "libs/gotrue-entity",
     "admin_frontend",
     "libs/app-error",
+    "libs/workspace-access",
     "libs/workspace-template",
     "libs/encrypt",
     "libs/access-control",
@@ -163,6 +165,7 @@ database-entity = { path = "libs/database-entity" }
 shared-entity = { path = "libs/shared-entity" }
 gotrue-entity = { path = "libs/gotrue-entity" }
 access-control = { path = "libs/access-control" }
+workspace-access = { path = "libs/workspace-access" }
 app-error = { path = "libs/app-error" }
 async-trait = "0.1.77"
 prometheus-client = "0.22.0"
