+++ services/appflowy-collaborate/src/collab/mod.rs
@@ -2,3 +2,4 @@ pub mod access_control;
 pub mod cache;
 pub mod disk_cache;
 pub mod mem_cache;
+pub mod notification;
