--- src/biz/casbin/mod.rs
@@ -1,4 +1,0 @@
-pub mod pg_listen;
-mod workspace_ac;
-
-pub use workspace_ac::WorkspaceAccessControlImpl;
