+++ src/biz/mod.rs
@@ -1,5 +1,4 @@
 pub mod actix_ws;
-pub mod casbin;
 pub mod chat;
 pub mod collab;
 pub mod pg_listener;
